move ./api to ./internal/api and upd refs
