+++ internal/api/entry.go
